moved table-cell back to main package
