+++ projects/ui-framework/bob-table/src/public_api.ts
@@ -12,10 +12,4 @@ export { ActionsCellComponent } from './table/table-cell-components/actions-cell
 export { GridActions } from './table/table-cell-components/actions-cell/actions-cell.interface';
 export { TableUtilsService } from './table/table-utils-service/table-utils.service';
 export { TableModule } from './table/table.module';
-// Card Table
-export { CardTableModule } from './table/card-table/card-table.module';
-export { CardTableComponent } from './table/card-table/card-table/card-table.component';
-export { CellWidthsService } from './table/card-table/cell-widths-service/cell-widths.service';
-export { TableCardCellComponent } from './table/card-table/table-card-cell/table-card-cell.component';
-export { TableCardComponent } from './table/card-table/table-card/table-card.component';
-export * from './table/card-table/card-table.interface';
+
