+++ projects/ui-framework/bob-table/src/table/table/styles/table.component.scss
@@ -1,4 +1,4 @@
-@import '~bob-style/style/common-imports';
+@import '../../../../../src/lib/style/common-imports';
 
 $min-table-height: 140px;
