+++ projects/ui-framework/src/lib/form-elements/datepicker/datepicker.component.spec.ts
@@ -203,4 +203,10 @@ describe('DatepickerComponent', () => {
       });
     });
   });
+
+  describe('Month picker', () => {
+    beforeEach(() => {});
+
+    it('should ', () => {});
+  });
 });
