+++ projects/ui-framework/src/lib/table/card-table/cell-widths-service/cell-widths.service.spec.ts
