+++ projects/ui-framework/src/lib/table/card-table/table-card-cell/table-card-cell-mixin.scss
