+++ projects/ui-framework/src/lib/table/card-table/table-card-cell/table-card-cell.component.scss
@@ -1,4 +1,4 @@
-@import '~bob-style/style/common-imports';
+@import '../../../style/common-imports';
 @import './table-card-cell-mixin';
 
 :host {
