+++ projects/ui-framework/src/lib/table/card-table/table-card/table-card.component.scss
